# optimize InterfaceBinderAnalyzer