--- a/tests/D2L.CodeStyle.Analyzers.Test/Specs/InterfaceBinderAnalyzer.cs
+++ b/tests/D2L.CodeStyle.Analyzers.Test/Specs/InterfaceBinderAnalyzer.cs
@@ -5,7 +5,21 @@ using D2L.LP.Serialization;
 
 namespace D2L.LP.LayeredArch.Data {
 
-	public static class InterfaceBinder<T> { }
+	public static class InterfaceBinder<T> {
+
+		private void PrivateMethod() {
+			// This would be an error if the analyzer didn't only consider the public API
+			PrivateMethod2();
+		}
+		private void PrivateMethod2() { }
+
+		public static object AllFields { get; } = null;
+		public static object CreateAllFields() => null;
+
+		public static object MatchedFields { get; } = null;
+		public static object CreateMathedFields() => null;
+
+	}
 }
 
 namespace SpecTests {
@@ -23,34 +37,36 @@ namespace SpecTests {
 	public class TestCases {
 
 		public void UsedWithClass() {
-			var type = typeof( InterfaceBinder< /* InterfaceBinder_InterfacesOnly(ClassDto) */ ClassDto /**/ > );
+			// This used to be considered an error; however, the updated version of the analyzer
+			// only looks at member accesses rather than generic type names to be less expensive
+			_ = typeof( InterfaceBinder<ClassDto> );
+
+			_ = /* InterfaceBinder_InterfacesOnly(ClassDto) */ InterfaceBinder<ClassDto>.AllFields /**/;
+			_ = /* InterfaceBinder_InterfacesOnly(ClassDto) */ InterfaceBinder<ClassDto>.CreateAllFields() /**/;
+			_ = /* InterfaceBinder_InterfacesOnly(ClassDto) */ InterfaceBinder<ClassDto>.MatchedFields /**/;
+			_ = /* InterfaceBinder_InterfacesOnly(ClassDto) */ InterfaceBinder<ClassDto>.CreateMathedFields() /**/;
 		}
 
 		public void UsedWithInterface() {
-			var type = typeof( InterfaceBinder<InterfaceDto> );
+			_ = typeof( InterfaceBinder<InterfaceDto> );
+
+			_ = InterfaceBinder<InterfaceDto>.AllFields;
+			_ = InterfaceBinder<InterfaceDto>.CreateAllFields;
+			_ = InterfaceBinder<InterfaceDto>.MatchedFields;
+			_ = InterfaceBinder<InterfaceDto>.CreateMathedFields;
 		}
 
 		public void GenericTypeDefinition() {
-			var type = typeof( InterfaceBinder<> );
+			_ = typeof( InterfaceBinder<> );
 		}
 
 		public void UnknownGenericTypeArgument() {
-			var type = typeof( InterfaceBinder< /* InterfaceBinder_InterfacesOnly(Wacky) */ Wacky /**/ > );
-		}
-	}
+			_ = typeof( InterfaceBinder<Wacky> );
 
-	public class Unrelated {
-
-		public void UnknownGenericTypeDefinition() {
-			var type = typeof( Wacky<int> );
-		}
-
-		public void NotInterfaceBinder() {
-			var type = typeof( Action<int> );
-		}
-
-		public void TooManyGenericTypeArguments() {
-			var type = typeof( Action<int, int> );
+			_ = /* InterfaceBinder_InterfacesOnly(Wacky) */ InterfaceBinder<Wacky>.AllFields /**/;
+			_ = /* InterfaceBinder_InterfacesOnly(Wacky) */ InterfaceBinder<Wacky>.CreateAllFields() /**/;
+			_ = /* InterfaceBinder_InterfacesOnly(Wacky) */ InterfaceBinder<Wacky>.MatchedFields /**/;
+			_ = /* InterfaceBinder_InterfacesOnly(Wacky) */ InterfaceBinder<Wacky>.CreateMathedFields() /**/;
 		}
 	}
 }