With ImmutabilityAnalyzer as reference (given run variability):

| Analyzer | Before | After |
| -- | -- | -- |
| ImmutabilityAnalyzer | 11.804s | 11.681s |
| InterfaceBinderAnalyzer | 5.898s | 342ms |